% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts1.R
\name{dts1}
\alias{dts1}
\title{Dataframe for ggplottimeseries}
\usage{
dts1(x, y, z, type = "additive")
}
\arguments{
\item{x}{a vector or matrix of the dates formatted as YYYY-MM-DD}

\item{y}{a vector or matrix of the observed time-series values}

\item{z}{frequency (the number of observations per unit of time)}

\item{type}{type of time series, either "additive" (default) or multiplicative}
}
\value{
This returns to a list containing all LME models arranged in the order specified by the argument xy
}
\description{
This function creates a data frame of decomposed time series.
}
\examples{
data(sst)

x <- sst$date
y <- sst$sst
z <- 365.25 #number of days in a year
df <- dts1(x,y,z, type = "additive")
head(df)

}
\references{
Garrett Grolemund, Hadley Wickham (2011). Dates and Times Made Easy with lubridate. Journal of Statistical Software, 40(3), 1-25. URL http://www.jstatsoft.org/v40/i03/.

Hyndman R, Athanasopoulos G, Bergmeir C, Caceres G, Chhay L, O'Hara-Wild M, Petropoulos F, Razbash S, Wang E, Yasmeen F (2018). _forecast: Forecasting functions for time series and linear models_. R package version 8.4, <URL: http://pkg.robjhyndman.com/forecast>.

Hyndman RJ, Khandakar Y (2008). “Automatic time series forecasting: the forecast package for R.”_Journal of Statistical Software_, *26*(3), 1-22. <URL: http://www.jstatsoft.org/article/view/v027i03>.
}
\author{
Brisneve Edullantes
}
