% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts2.R
\name{dts2}
\alias{dts2}
\title{Dataframe for ggplottimeseries}
\usage{
dts2(x, type = "additive")
}
\arguments{
\item{x}{a time series-class data}

\item{type}{type of time series, either "additive" (default) or multiplicative}
}
\value{
This returns to a dataframe with the components of timeseries
\item{date}{date formatted as YYYY-MM-DD}
\item{observation}{observation of the timeseries}
\item{trend}{trend component of the timeseries}
\item{seasonal}{seasonal component of the timeseries}
\item{random}{random component of the timeseries}
}
\description{
This function converts time series-class data into a data frame of decomposed time series.
}
\examples{
data(co2)
df <- dts2(co2, type ="additive")
head(df)

}
\author{
Brisneve Edullantes
}
