% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdecompose.R
\name{ggdecompose}
\alias{ggdecompose}
\alias{ggobserve}
\alias{ggtrend}
\alias{ggseason}
\alias{ggrandom}
\alias{ggdetrend}
\alias{ggdeseason}
\title{Visualisation of decomposed time series}
\usage{
ggdecompose(x)
ggobserve(x)
ggtrend(x)
ggseason(x)
ggrandom(x)
ggdetrend(x)
ggdeseason(x)
}
\arguments{
\item{x}{a data frame generated by either dts or dts2 functions.}
}
\value{
This returns to a plot.
}
\description{
These functions plots the observed, trend, seasonal, and random components of time series
into one figure (ggdecompose) or into separate figures (ggobserve, ggtrend, ggseason, ggrandom,
respectively). These functions also plots detrended and deseasonalised time series
(ggdetrend and ggdeseason, respectively). These can be integrated with ggplot functions.
}
\examples{
data(co2)
x <- dts2(co2, type ="additive")
#plots decomposed time series into one figure
ggdecompose(x)+
 xlab("Date")+
 ylab("Atmospheric Concentration of CO2")
#plots components of time series into separate figure
ggobserve(x)+
 xlab("Date")+
 ylab("Observed Atmospheric Concentration of CO2")

ggtrend(x)+
 xlab("Date")+
 ylab("Trend of Atmospheric Concentration of CO2")

ggseason(x)+
 xlab("Date")+
 ylab("Seasonality of Atmospheric Concentration of CO2")

ggrandom(x)+
 xlab("Date")+
 ylab("Random Variation of Atmospheric Concentration of CO2")

#plots detrended and deseasonalised Time Series

ggdetrend(x)

ggdeseason(x)

}
\author{
Brisneve Edullantes
}
