% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\docType{data}
\name{sst}
\alias{sst}
\title{Sea surface temperature data}
\format{a sample dataframe with 2142 observations on the following variables.
\describe{
  \item{\code{date}}{a date vector of the time of the observation}
  \item{\code{sst}}{a numeric vector}}}
\description{
Daily sea surface temperature data downloaded from Giovanni
}
\author{
Brisneve Edullantes
}
